// src/data/confirmationHd.type.ts
/*
Author: Giovanni Ranieri
Year: 2026
Description: Confirmation HD custom message to be displayed
*/

export type ConfirmationHd = {
	default: boolean;
	title: string;
    left_color: string;
    right_color: string;
    left_text: string;
    right_text: string;
    text_color_left: string;
    text_color_right: string;
};